Antenna files
The NXP design doc AN1445 (Antenna design for MFRC522) contains
the recommended impedance matching circuitry and calculations. NXP
makes an excel spreadsheet available to make the calculations a
little easier, but I refuse to use excel so here we are.

I made a matlab script to do the calc necessar for the HF impedance
matching.